Handle combined additional attributes
When an additional attribute in the metadata is a result of combining
metadata from a split granule, it results in an attribute value that is
a combination of the separate metadata, where the 2 separate values are
joined with `" + "`. In this case set the attribute value to the value
preceding the `" + "`.